refactor: add sub-folders for some structure
