+++ package.json
@@ -23,7 +23,7 @@
     },
     "type": "module",
     "scripts": {
-        "build": "tsc",
+        "build": "shx rm -r dist/ && tsc",
         "start": "npm run build && npm run start:prod",
         "start:prod": "node ./dist/src/main.js",
         "test": "jest",
@@ -44,6 +44,7 @@
     "devDependencies": {
         "@types/jshint": "^2.12.1",
         "prettier": "^2.8.4",
+        "shx": "^0.3.4",
         "ts-jest": "^29.0.5",
         "typescript": "^4.9.5"
     }
