+++ src/command/compute-crap.command.ts
@@ -1,9 +1,9 @@
 import { Logger } from "@nestjs/common";
 import { JSHINT } from "jshint";
 import { CommandRunner, InquirerService, RootCommand } from "nest-commander";
-import { crap } from "./crap-score.js";
-import { FileSystemService } from "./file-system.service.js";
-import { getCoverageForFunction } from "./function-coverage.js";
+import { crap } from "../computation/crap-score.js";
+import { getCoverageForFunction } from "../computation/function-coverage.js";
+import { FileSystemService } from "../file-system.service.js";
 
 @RootCommand({
     arguments: "[testCoveragePath]",
