+++ src/computation/crap-score.ts
